% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckGeneSymbol.r
\name{.CheckGeneSymbol}
\alias{.CheckGeneSymbol}
\title{Check DepOIs in the fingerprint database}
\usage{
.CheckGeneSymbol(dep.data, filename.out)
}
\description{
Check DepOIs in the fingerprint database
}
